requirements.txt
numpy
flask
flask-cors
googletrans==3.1.0a0
opencv-python
pillow
pytesseract
requests
waitress